+++ next.config.mjs
@@ -1,8 +1,16 @@
 /** @type {import('next').NextConfig} */
+/** @typedef {import('./types/next-config').Redirect} Redirect */
+
+/**
+ * @type {import('next').NextConfig}
+ */
 const nextConfig = {
   images: {
     unoptimized: true,
   },
+  /**
+   * @returns {Promise<Redirect[]>}
+   */
   async redirects() {
     return [
       { source: "/app", destination: "/app/flux", permanent: false },